automation/requirements.txt
# Morning Brief Automation - Python Dependencies

# For GitHub API
requests>=2.31.0

# For Gmail API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# For timezone handling
pytz>=2023.3
